django/requirements.txt
wheel
psycopg2-binary
Django
django-simple-history
#https://pypi.org/project/django-redis/
#https://pypi.org/project/django-select2/
django-select2
django-admin-ordering
django-colorfield
django-json-widget
django-cors-headers
djangorestframework
Markdown
django-filter
django-rest-auth
django-rest-registration
django-postgres-createdb
drf-flex-fields
drf-renderer-xlsx
drf-writable-nested
drf-yasg
celery
django-celery-beat
django-celery-results
django-extensions
preview-generator
reportlab
#https://pypi.org/project/reportlab-qrcode/
reportlab-qrcode
#https://textract.readthedocs.io/en/stable/installation.html
##textract==1.6.3
#https://pypi.org/project/docxtpl/
docxtpl
#https://pypi.org/project/Babel/
Babel
#https://pypi.org/project/num2words/
num2words
#https://pypi.org/project/filehash/
filehash
#https://pypi.org/project/zeep/
zeep
#https://pypi.org/project/declxml/
declxml
#https://pycallgraph.readthedocs.io/en/master/
##pycallgraph
#django-fixture-magic==0.1.5
gunicorn
# ujson
# orjson
django-cacheops
django-redis
django-cachalot
PyPDF2
channels
channels-redis
pandas
django-jazzmin
django-subadmin
pysftp
XlsxWriter
beautifulsoup4
python-logstash
JSON-log-formatter
viberbot
django-grappelli
django-debug-toolbar
